refactor(SR-8): separate app creation for reuse in portal-api
